requirements.txt
openshift==0.12.0
PyYAML==6.0
